+++ src/command/mod.rs
@@ -11,6 +11,13 @@ pub struct SimpleGit {
 pub enum GitSubCommand {
     #[command(name = "init", about = "init a git repo")]
     Init(InitOpts),
+    #[command(name = "add", about = "add files into index")]
+    Add(AddOpts),
+    #[command(
+        name = "commit",
+        about = "Create a new commit containing the current contents of the index and the given log message describing the changes"
+    )]
+    Commit(CommitOpts),
 }
 
 #[derive(Debug, Parser)]
@@ -33,4 +40,38 @@ pub struct InitOpts {
     /// Use <branch-name> for the initial branch in the newly created repository. If not specified, fall back to the default name.
     #[arg(short, long = "initial-branch")]
     pub branch: String,
+}
+
+#[derive(Debug, Parser)]
+pub struct AddOpts {
+    /// Files to add content from
+    #[arg(short, long = "path")]
+    pub path_spec: String,
+
+    /// If no <pathspec> is given when -A option is used, all files in the entire working tree are updated
+    #[arg(short, long = "all")]
+    pub all: bool,
+}
+
+#[derive(Debug, Parser)]
+pub struct CommitOpts {
+    /// Tell the command to automatically stage files that have been modified and deleted, but new files you have not told Git about are not affected.
+    #[arg(short, long)]
+    pub all: bool,
+
+    /// Record changes to the repository
+    #[arg(short, long)]
+    pub message: String,
+
+    /// The message taken from file with -F, command line with -m, and from commit object with -C are usually used as the commit log message unmodified. This option lets you further edit the message taken from these sources.
+    #[arg(short, long = "edit")]
+    pub edit: bool,
+
+    /// Use the selected commit message without launching an editor
+    #[arg(short, long = "no-edit")]
+    pub no_edit: bool,
+
+    /// Replace the tip of the current branch by creating a new commit.
+    #[arg(long)]
+    pub amend: bool,
 }